backend-api/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-cors==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
prometheus-client==0.20.0